# Conflicts:
#	src/cmdmapper.h
#	src/commentscan.l
#	src/docparser.cpp
#	src/doctokenizer.h
#	src/doctokenizer.l
#	src/markdown.cpp
#	src/pyscanner.l

+++ src/htags.cpp
@@ -160,8 +160,8 @@ QCString Htags::path2URL(const QCString &path)
 {
   QCString url,symName=path;
   QCString dir = g_inputDir.absPath();
-  int dl=dir.length();
-  if ((int)symName.length()>dl+1)
+  size_t dl=dir.length();
+  if (symName.length()>dl+1)
   {
     symName = symName.mid(dl+1);
   }
